Handle cases where keymap from emojify are invoked outside an emoji
This happens if point is immediately after an emoji, in which case the
keymap property is inherited from the preceding emoji. This could have
been disabled using the 'rear-nonsticky' text property but that breaks
the backspaces keybinding, since point will then need to be 'inside' an
emoji for emojify's keybinding to be invoked. As such backspace
immediately after an emoji simply deletes the preceding char, unemojifing
the text (thus removing the emojify keymap as well) making the backspace
key bind added by emojify useless.

The current simply invokes whatever non-emojify keybinding is active at
the point