Makes comint derived modes act like prog derived
Makes comint-mode derived modes act as if they are prog-mode
derived. This means instead of having emojify disabled by default in
these modes, unicode emojis are emojified and whatever rules the user
has set for programming based rules applies to these modes as
well. Since comint mode is described as being for a
"command-interpreter-in-a-buffer" which is essentially a repl, it makes
sense to treat it as a programming mode. This would allow shell-mode to
be able to display emoji in a (hopefully) sensible way.