Do not scale created images
Starting from Emacs 26.1, images are automatically scaled to match the
frame DPI value (HiDPI support), unless `image-scaling-factor' is set
to a value different than auto. However, the height of the emoji
images is computed from the height of the current font. Therefore,
there is no need to scale them. Therefore, we enforce a scale of 1.